Fix bed render crash

+++ DwarfCorp/DwarfCorpXNA/Gui/RenderData.cs
@@ -43,7 +43,7 @@ namespace DwarfCorp.Gui
             float scaleX = ActualScreenBounds.X/1920.0f;
             float scaleY = ActualScreenBounds.Y/1080.0f;
             float maxScale = Math.Max(scaleX, scaleY);
-            var scale = (int) MathFunctions.Clamp((int)Math.Ceiling(maxScale), 1, 10);
+            var scale = MathFunctions.Clamp((int)Math.Ceiling(maxScale), 1, 10);
             GameSettings.Default.GuiScale = scale;
             return scale;
         }
